部分修改
1)有些不需要加锁的地方 去掉
2)加入log

+++ meta_header/netstream.h
@@ -189,7 +189,7 @@ public:
 
 	bool WriteFloat(float fData)
 	{
-		int nData = fData * 256;
+		int nData = (int)(fData * 256);
 		return WriteInt(nData);
 	}
